+++ lending/src/funcs/accountsPayableBillsList.ts
@@ -116,6 +116,7 @@ export async function accountsPayableBillsList(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "GET",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     query: query,