+++ lending/src/funcs/accountsReceivableInvoicesDownloadPdf.ts
@@ -109,6 +109,7 @@ export async function accountsReceivableInvoicesDownloadPdf(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "GET",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     body: body,