+++ lending/src/funcs/bankStatementsStartUploadSession.ts
@@ -116,6 +116,7 @@ export async function bankStatementsStartUploadSession(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "POST",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     body: body,