+++ lending/src/funcs/bankingAccountsGet.ts
@@ -114,6 +114,7 @@ export async function bankingAccountsGet(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "GET",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     body: body,