+++ lending/src/funcs/companiesDelete.ts
@@ -103,6 +103,7 @@ export async function companiesDelete(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "DELETE",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     body: body,