+++ lending/src/funcs/connectionsGet.ts
@@ -106,6 +106,7 @@ export async function connectionsGet(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "GET",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     body: body,