+++ lending/src/funcs/connectionsUnlink.ts
@@ -107,6 +107,7 @@ export async function connectionsUnlink(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "PATCH",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     body: body,