+++ lending/src/funcs/liabilitiesGenerateLoanSummary.ts
@@ -111,6 +111,7 @@ export async function liabilitiesGenerateLoanSummary(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "POST",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     query: query,