+++ lending/src/funcs/loanWritebackSuppliersGetCreateUpdateModel.ts
@@ -115,6 +115,7 @@ export async function loanWritebackSuppliersGetCreateUpdateModel(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "GET",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     body: body,