+++ lending/src/funcs/loanWritebackTransfersCreate.ts
@@ -120,6 +120,7 @@ export async function loanWritebackTransfersCreate(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "POST",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     query: query,