+++ lending/src/funcs/manageDataRefreshAllDataTypes.ts
@@ -105,6 +105,7 @@ export async function manageDataRefreshAllDataTypes(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "POST",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     body: body,