+++ lending/src/funcs/manageDataRefreshDataType.ts
@@ -112,6 +112,7 @@ export async function manageDataRefreshDataType(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "POST",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     query: query,