+++ lending/src/funcs/salesMetricsGetCustomerRetention.ts
@@ -144,6 +144,7 @@ export async function salesMetricsGetCustomerRetention(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "GET",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     query: query,