+++ lending/src/funcs/salesMetricsGetLifetimeValue.ts
@@ -136,6 +136,7 @@ export async function salesMetricsGetLifetimeValue(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "GET",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     query: query,