+++ lending/src/funcs/transactionsAccountTransactionsGet.ts
@@ -118,6 +118,7 @@ export async function transactionsAccountTransactionsGet(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "GET",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     body: body,