+++ lending/src/funcs/transactionsTransfersList.ts
@@ -118,6 +118,7 @@ export async function transactionsTransfersList(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "GET",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     query: query,